+++ src/test/integration/build_test.ts
@@ -142,6 +142,24 @@ suite('polymer build', function() {
         path.join(fixturePath, 'polymer-2-project', 'expected/default'));
   });
 
+  test('handles bundle tagged-template literals in ES5', async () => {
+    const tmpDir = tmp.dirSync();
+    copyDir(path.join(fixturePath, 'build-with-tagged-template-literals', 'source'), tmpDir.name);
+
+    await runCommand(binPath, ['build'], {
+      cwd: tmpDir.name,
+    });
+
+    const filename = path.join(tmpDir.name, 'build', 'es5-bundled', 'my-app.html');
+    const contents = fs.readFileSync(filename, 'utf-8');
+    // assert contents contain _templateObject with UUID suffix
+    assert.match(contents, /_templateObject\d*_[A-Fa-f0-9]+\s*=/g,
+                 'build output does not contain modified _templateObject names');
+    // assert contents don't contain unmodified "_templateObject" variable
+    assert.notMatch(contents, /_templateObject\d*\s*=/g,
+                 'build output contains unmodified _templateObject names');
+  });
+
   test('--npm finds dependencies in "node_modules/"', async () => {
     const tmpDir = tmp.dirSync();
     copyDir(path.join(fixturePath, 'element-with-npm-deps'), tmpDir.name);
@@ -159,6 +177,7 @@ suite('polymer build', function() {
           cwd: tmpDir.name
         });
       });
+
 });
 
 function copyDir(fromDir: string, toDir: string) {